Add basic statsd, archive log, and @logstash_uuid
@logstash_uuid allows us to reindex logs into elasticsearch. The
use of document_id on the elasticsearch output forces it to write
the internal _id to be the same as @logstash_uuid.

This in turn ensures that a message with the same @logstash_uuid is
replaced on reinsertion into ES - so we can rejig fields, reindex the
database, etc.

Creating an archive log (with the file output), and adding a
corresponding listening unix socket input, gives us the ability to do
just this -- as well as a convenient 'output archive'.

The statsd output is there to see how quickly we're processing messages.