Add tests for 200_filter_all config
Many of the tests in syslog_input_spec.rb do not need to be
tested via the syslog input - we just need to ensure that
the data provided to a filter-only config object is as it would
be if the syslog filter had parsed the original event.

This sets up the structure for moving most of the syslog_input_spec
tests out of there, and just ensuring that the syslog input is
doing what we expect.