Simplify type=>syslog tests via a direct filter test suite
The hacking to get a syslog input handler in place is excessive for
most of our tests, when all we need to know is that the filter configuration
has received the message from the input in the right form.

Migrate the syslog_input_test test cases into a specific test based
around the 200_filter_all config.