Stop logstash-web from starting.
The logstash web process was chewing up lots of CPU, due to a bad
cmdline option (-l) in the upstart config.

This basically mean that the JVM kept trying to start, taking up resource.

We do not use logstash-web, as Kibana does what we need, so safest thing is to
just not start it.