Prevent cleaning of conf.d dir from causing reload
Finally worked this out. In order to clean a directory out but
leave files there that are needed, you must *require* the files
in the `file.directory`.

This however creates a chicken-vs-egg problem, as you are effectively
requiring a file to be present before the directory.

To hack around this, we add the 'makedirs: True' option to our files,
which removes this limitation.

This is documented in https://github.com/saltstack/salt/issues/12965

NB: In this particular instance it makes more sense for us to `require_in`
the clean directory in each file. It has the same effect though.