Discovery that timestamp.iso8601 is broken
Since the shift in year to 2015, these tests are failing. Added a
few more assertations on timestamp.iso8601 to see the scope of this
issue.

Originally I thought this was a problem with the date{} filter, but
this is not used for the syslog parsing AIUI, so leaving these extra
(failing) tests in here for now.