Fix _grokparsefailure for mongo 'command' ops.
The 'command' mongo operation has a different format message for
slow query logs, causing a _grokparsefailure. (the actual failure was due to
it operating on the '$cmd' collection, but the message is generally quite different too)

This commit fixes the $cmd collection problem, so we should now lose the failures.